+++ gcovr/formats/html/write.py
@@ -17,22 +17,13 @@
 #
 # ****************************************************************************
 
-import logging
-import os
 import hashlib
 import io
+import logging
+import os
 from argparse import ArgumentTypeError
 from typing import Any, Callable, Dict, Optional, Union
 
-from ...options import Options
-
-from ...version import __version__
-from ...utils import (
-    force_unix_separator,
-    realpath,
-    commonpath,
-    open_text_for_writing,
-)
 from ...coverage import (
     CallCoverage,
     CovData,
@@ -48,6 +39,14 @@ from ...coverage import (
     SummarizedStats,
     sort_coverage,
 )
+from ...options import Options
+from ...utils import (
+    commonpath,
+    force_unix_separator,
+    open_text_for_writing,
+    realpath,
+)
+from ...version import __version__
 
 LOGGER = logging.getLogger("gcovr")
 PYGMENTS_CSS_MARKER = "/* Comment.Preproc */"
@@ -70,15 +69,36 @@ class Lazy:
         return self.get(*args)
 
 
-# Loading Jinja and preparing the environmen is fairly costly.
+# html_theme string is <theme_directory>.<color> or only <color> (if only color use default)
+# examples: github.green github.blue or blue or green
+def get_theme_name(html_theme: str) -> str:
+    return html_theme.split(".")[0] if "." in html_theme else "default"
+
+
+def get_theme_color(html_theme: str) -> str:
+    return html_theme.split(".")[1] if "." in html_theme else html_theme
+
+
+# Loading Jinja and preparing the environment is fairly costly.
 # Only do this work if templates are actually used.
 # This speeds up text and XML output.
 @Lazy
-def templates():
-    from jinja2 import Environment, PackageLoader
+def templates(options):
+
+    from jinja2 import ChoiceLoader, Environment, FileSystemLoader, PackageLoader
+
+    # As default use the package loader
+    loader = PackageLoader(
+        "gcovr.formats.html",
+        package_path=get_theme_name(options.html_theme),
+    )
+
+    # If a directory is given files in the directory have higher precedence.
+    if options.html_template_dir is not None:
+        loader = ChoiceLoader([FileSystemLoader(options.html_template_dir), loader])
 
     return Environment(
-        loader=PackageLoader("gcovr.formats.html"),
+        loader=loader,
         autoescape=True,
         trim_blocks=True,
         lstrip_blocks=True,
@@ -110,7 +130,14 @@ def user_templates():
 
 class CssRenderer:
 
-    THEMES = ["green", "blue"]
+    THEMES = [
+        "green",
+        "blue",
+        "github.blue",
+        "github.green",
+        "github.dark-green",
+        "github.dark-blue",
+    ]
 
     @staticmethod
     def get_themes():
@@ -126,7 +153,7 @@ class CssRenderer:
             template_path = os.path.relpath(options.html_css)
             return user_templates().get_template(template_path)
 
-        return templates().get_template("style.css")
+        return templates(options).get_template("style.css")
 
     @staticmethod
     def render(options):
@@ -146,12 +173,12 @@ class NullHighlighting:
 
 
 class PygmentHighlighting:
-    def __init__(self):
+    def __init__(self, style: str):
         self.formatter = None
         try:
             from pygments.formatters.html import HtmlFormatter
 
-            self.formatter = HtmlFormatter(nowrap=True)
+            self.formatter = HtmlFormatter(nowrap=True, style=style)
         except ImportError as e:  # pragma: no cover
             LOGGER.warning(f"No syntax highlighting available: {str(e)}")
 
@@ -168,8 +195,8 @@ class PygmentHighlighting:
             return NullHighlighting.highlighter_for_file(filename)
 
         import pygments
-        from pygments.lexers import get_lexer_for_filename
         from markupsafe import Markup
+        from pygments.lexers import get_lexer_for_filename
 
         try:
             lexer = get_lexer_for_filename(filename, None, stripnl=False)
@@ -183,8 +210,13 @@ class PygmentHighlighting:
 
 @Lazy
 def get_formatter(options):
+    highlight_style = (
+        templates(options)
+        .get_template(f"pygments.{get_theme_color(options.html_theme)}")
+        .render()
+    )
     return (
-        PygmentHighlighting()
+        PygmentHighlighting(highlight_style)
         if options.html_syntax_highlighting
         else NullHighlighting()
     )
@@ -390,7 +422,7 @@ def write_report(covdata: CovData, output_file: str, options: Options) -> None:
             css_link = css_output
         data["css_link"] = css_link
 
-    data["theme"] = options.html_theme
+    data["theme"] = get_theme_color(options.html_theme)
 
     root_info.set_coverage(covdata)
 
@@ -400,10 +432,10 @@ def write_report(covdata: CovData, output_file: str, options: Options) -> None:
     filtered_fname = ""
     sorted_keys = sort_coverage(
         covdata,
-        show_branch=False,
-        filename_uses_relative_pathname=True,
+        by_branch=False,
         by_num_uncovered=options.sort_uncovered,
         by_percent_uncovered=options.sort_percent,
+        filename_uses_relative_pathname=True,
     )
 
     if options.html_nested:
@@ -480,7 +512,7 @@ def write_root_page(output_file: str, options, data: Dict[str, Any]) -> None:
     #
     # Generate the root HTML file that contains the high level report
     #
-    html_string = templates().get_template("directory_page.html").render(**data)
+    html_string = templates(options).get_template("directory_page.html").render(**data)
     with open_text_for_writing(
         output_file, encoding=options.html_encoding, errors="xmlcharrefreplace"
     ) as fh:
@@ -593,7 +625,7 @@ def write_source_pages(
             error_no_files_not_found += 1
         os.chdir(current_dir)
 
-        html_string = templates().get_template("source_page.html").render(**data)
+        html_string = templates(options).get_template("source_page.html").render(**data)
         with open_text_for_writing(
             cdata_sourcefile[f],
             encoding=options.html_encoding,
@@ -602,7 +634,7 @@ def write_source_pages(
             fh.write(html_string + "\n")
 
     data["all_functions"] = [all_functions[k] for k in sorted(all_functions)]
-    html_string = templates().get_template("functions_page.html").render(**data)
+    html_string = templates(options).get_template("functions_page.html").render(**data)
     with open_text_for_writing(
         functions_output_file,
         encoding=options.html_encoding,
@@ -639,10 +671,10 @@ def write_directory_pages(
 
         sorted_files = sort_coverage(
             directory.children,
-            show_branch=False,
-            filename_uses_relative_pathname=True,
+            by_branch=False,
             by_num_uncovered=options.sort_uncovered,
             by_percent_uncovered=options.sort_percent,
+            filename_uses_relative_pathname=True,
         )
 
         root_info.clear_files()
@@ -653,7 +685,9 @@ def write_directory_pages(
                 directory.children[key], cdata_sourcefile[fname], cdata_fname[fname]
             )
 
-        html_string = templates().get_template("directory_page.html").render(**data)
+        html_string = (
+            templates(options).get_template("directory_page.html").render(**data)
+        )
         filename = None
         if f in [root_key, ""]:
             filename = output_file