+++ gcovr/formats/json/write.py
@@ -98,7 +98,7 @@ def write_summary_report(covdata, output_file: str, options: Options):
     # Data
     keys = sort_coverage(
         covdata,
-        show_branch=options.show_branch,
+        by_branch=options.show_branch,
         by_num_uncovered=options.sort_uncovered,
         by_percent_uncovered=options.sort_percent,
     )
@@ -232,6 +232,8 @@ def _json_from_function(function: FunctionCoverage) -> list:
             "name": function.name,
             "lineno": lineno,
             "execution_count": count,
+            "returned_count": function.returned[lineno],
+            "blocks_percent": function.blocks[lineno],
         }
         if function.excluded[lineno]:
             json_function["gcovr/excluded"] = True