+++ gcovr/tests/decisions/switch_test.cpp
@@ -1,0 +1,10 @@
+
+#if (defined __GNUC__ && (__GNUC__ >= 6)) || (defined __clang_major__)
+#include "switch_test.h"
+
+SwitchTestIssue783::SwitchTestIssue783() : _name("test") {}
+
+void SwitchTestIssue783::checkSwitch() {
+    doSomething(SomeEnum::EnumValue_1);
+}
+#endif